+++ sealed_annotations/lib/sealed_annotations.dart
@@ -2,6 +2,5 @@
 library sealed_annotations;
 
 export 'package:equatable/equatable.dart';
-export 'package:meta/meta.dart';
 
 export 'src/annotations.dart';