+++ sealed_generators/lib/src/manifest/manifest_reader.dart
@@ -1,13 +1,10 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
-import 'package:meta/meta.dart';
-import 'package:sealed_annotations/sealed_annotations.dart';
 import 'package:sealed_generators/src/manifest/annotation_utils.dart';
 import 'package:sealed_writer/sealed_writer.dart';
 
 /// manifest reader built by manifest reader builder
-@sealed
 class ManifestReader {
   ManifestReader({
     required this.topName,