+++ sealed_writer/lib/src/writer/base/base_cast_utils_writer.dart
@@ -1,9 +1,7 @@
-import 'package:meta/meta.dart';
 import 'package:sealed_writer/src/manifest/manifest.dart';
 import 'package:sealed_writer/src/writer/base/base_utils_writer.dart';
 
 /// top cast writer
-@immutable
 abstract class BaseCastUtilsWriter extends BaseUtilsWriter {
   const BaseCastUtilsWriter(Manifest manifest) : super(manifest);