+++ sealed_writer/lib/src/writer/sub/sub_field_writer.dart
@@ -1,11 +1,8 @@
-import 'package:meta/meta.dart';
 import 'package:sealed_writer/src/manifest/manifest.dart';
 import 'package:sealed_writer/src/utils/string_utils.dart';
 import 'package:sealed_writer/src/writer/base/base_utils_writer.dart';
 
 /// source writer
-@sealed
-@immutable
 class SubFieldWriter extends BaseUtilsWriter {
   const SubFieldWriter(Manifest manifest) : super(manifest);